Move feedback form to own page (#683)
