+++ apps/web/src/components/ui/label.tsx
@@ -6,12 +6,30 @@ import { cn } from "@/utils/cn";
 
 const labelVariants = cva(
   "text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70",
+  {
+    variants: {
+      intent: {
+        required: "content",
+      },
+    },
+  },
 );
 
-export const Label = React.forwardRef<
-  React.ElementRef<typeof LabelPrimitive.Root>,
-  React.ComponentPropsWithoutRef<typeof LabelPrimitive.Root> & VariantProps<typeof labelVariants>
->(({ className, ...props }, ref) => (
-  <LabelPrimitive.Root ref={ref} className={cn(labelVariants(), className)} {...props} />
-));
+type LabelProps = React.ComponentPropsWithoutRef<typeof LabelPrimitive.Root> &
+  VariantProps<typeof labelVariants> & {
+    required?: boolean;
+  };
+
+export const Label = React.forwardRef<React.ElementRef<typeof LabelPrimitive.Root>, LabelProps>(
+  ({ className, children, required = false, ...props }, ref) => (
+    <LabelPrimitive.Root ref={ref} className={cn(labelVariants(), className)} {...props}>
+      <span>{children}</span>
+      {required && (
+        <span className="ml-1 text-sm text-red-500" aria-hidden="true">
+          *
+        </span>
+      )}
+    </LabelPrimitive.Root>
+  ),
+);
 Label.displayName = LabelPrimitive.Root.displayName;
