+++ playwright/web/feedback.spec.ts
@@ -1,9 +1,8 @@
 import { expect, test } from "@playwright/test";
 
 test("test", async ({ page }) => {
-  await page.goto("/");
+  await page.goto("/tilbakemelding");
 
-  await page.getByTestId("feedback-button").click();
   await page.getByPlaceholder("Din tilbakemelding").click();
   await page.getByPlaceholder("Din tilbakemelding").fill("Hello world!");
